Add OLD to the DELETE notification
This way we can remove the event for example from the list in the frontend.
Without this information, we only "DELETE " without information about what is deleted.
Another issu, is when deleting many row we get only one event in the frontend. With this fix we get all deleted rows events.

NB : May Postgres dont send notification with same message many times.